+++ scripts/bench/CHANGELOG.md
@@ -8,7 +8,7 @@
   - @docknetwork/cheqd-blockchain-api@0.35.0
   - @docknetwork/credential-sdk@0.49.0
   - @docknetwork/cheqd-blockchain-modules@0.32.1
-  - @docknetwork/dock-blockchain-modules@0.28.1
+  - @docknetwork/dock-blockchain-modules@0.29.0
 
 ## 0.1.1